feat(recherche): recherche du site en popup superposé, avec historique et flou du contenu - SearchOverlay (Dialog MUI) ouvert depuis le header, état synchronisé   dans l'URL (?search=/&sq=/&sscope=) pour le retour arrière et les   liens profonds - Fix: SecondaryNavSm plantait sur les pages hors content/pages
