+++ src/components/_layout/AppBar/TopAppBar.jsx
@@ -9,6 +9,7 @@ import SideNavContent from '@/components/_layout/SideNav/SideNavContent'
 import MenuBurger from './MenuBurger'
 import LogoLink from './LogoLink'
 import SearchButton from './SearchButton'
+import SearchOverlay, { isSearchOverlayOpenInUrl } from './SearchOverlay'
 import pages from './menu'
 import { useLocation } from '@reach/router'
 
@@ -55,6 +56,7 @@ export default function TopAppBar({ lvl, location: propLocation = {} }) {
   const gatsbyLocation = useLocation()
   const location = propLocation || gatsbyLocation
   const [open, setOpen] = useState(false)
+  const [searchOpen, setSearchOpen] = useState(isSearchOverlayOpenInUrl) // ouverture de la modale de recherche (SearchOverlay)
   const theme = useTheme()
   const trigger = useScrollTrigger({
     disableHysteresis: true,
@@ -155,7 +157,7 @@ export default function TopAppBar({ lvl, location: propLocation = {} }) {
               )
             })}
             {/* Icône de recherche du site, avant le bouton "Je donne" */}
-            <SearchButton dark={trigger} />
+            <SearchButton dark={trigger} onClick={() => setSearchOpen(true)} />
             <JeDonneButton trigger={trigger} transitionProps={transitionProps} />
           </Stack>
           <Box >
@@ -170,6 +172,7 @@ export default function TopAppBar({ lvl, location: propLocation = {} }) {
       >
         <SideNavContent onClose={toggleDrawer(false)} />
       </SideNav>
+      <SearchOverlay open={searchOpen} onClose={() => setSearchOpen(false)} />
     </>
   )
 }
