+++ src/components/_layout/AppBar/TopAppBarSm.jsx
@@ -13,6 +13,7 @@ import SideNavSm from '@/components/_layout/SideNav/SideNavSm'
 import SideNavContent from '@/components/_layout/SideNav/SideNavContent'
 import MenuBurger from './MenuBurger'
 import SearchButton from './SearchButton'
+import SearchOverlay, { isSearchOverlayOpenInUrl } from './SearchOverlay'
 import LogoUdeMMonochrome from '@/images/logo-udem/logo_udem-noir.svg'
 
 const Offset = styled('div')(({ theme }) => theme.mixins.toolbar)
@@ -20,6 +21,7 @@ export const appBarHeight = '5rem'
 
 export default function TopAppBarSm(props) {
   const [open, setOpen] = useState(false)
+  const [searchOpen, setSearchOpen] = useState(isSearchOverlayOpenInUrl) // ouverture de la modale de recherche (SearchOverlay)
   const theme = useTheme()
   const isSmallScreen = useMediaQuery(theme.breakpoints.down('md'))
 
@@ -88,7 +90,7 @@ export default function TopAppBarSm(props) {
             <Box sx={{ display: 'flex', alignItems: 'center', ml: open ? 'auto' : 0 }}>
               {/* Fond du header toujours blanc ici : icône sombre pour rester visible
                   (contrairement à TopAppBar.jsx, blanche par défaut sur fond transparent) */}
-              {!open && <SearchButton dark />}
+              {!open && <SearchButton dark onClick={() => setSearchOpen(true)} />}
 
               <IconButton
                 onClick={toggleDrawer(true)}
@@ -111,6 +113,7 @@ export default function TopAppBarSm(props) {
       >
         <SideNavContent onClose={toggleDrawer(false)} />
       </SideNavSm>
+      <SearchOverlay open={searchOpen} onClose={() => setSearchOpen(false)} />
     </>
   )
 }
